requirements.txt
flask
bcrypt
pymongo
uuid
pyjwt